+++ src/FileTagSystem.h
@@ -1,8 +1,8 @@
 #ifndef FILE_TAG_SYSTEM_H
 #define FILE_TAG_SYSTEM_H
 
-#include "tag_manager.h"
-#include "user_manager.h"
+#include "TagManager.h"
+#include "UserManager.h"
 #include <string>
 #include <vector>
