+++ src/TagManager.h
